+++ src/components/MemoryGame/MemoryGame.js
@@ -6,16 +6,16 @@ import {GameControl} from './GameControl/GameControl';
 import {Gameboard} from './Gameboard/Gameboard';
 import styles from './MemoryGame.scss';
 import {Rules} from '../../shared/components/GameRules/Rules.js';
+import {isGameComplete} from './utils'
 
 export class MemoryGame extends React.Component {
   render() {
-    const isGameOver = !this.props.cards.find(c => !c.isMatched);
     return (
       <div className={styles.container} >
         <div className={styles.rulesContainer}>
           <Rules content={gameDescription()} show={this.props.showRules} onClick={() => this.props.appActions.memoryToggleRules()}/>   
         </div>
-        { isGameOver &&
+        { isGameComplete(this.props.cards)  &&
           <div className={styles.gameOver}>
             GAME OVER!
           </div>
