+++ src/components/MemoryGame/utils.js
@@ -1,6 +1,10 @@
 import {ICON_COUNT, iconNames} from './constants';
 import {randomArray} from '../../shared/utils';
 
+export const isGameComplete = (cards) => !cards.find(c => !c.isMatched); 
+export const randomCardArray = () => randomArray(ICON_COUNT * 2);
+export const hasUnmatched = (cards) => cards.filter((c) => c.isFlipped && !c.isMatched).length === 2;
+
 // build the cards array. 
 export const createCards = (randomArray) => {
   let cards = [];
@@ -17,10 +21,6 @@ export const createCards = (randomArray) => {
 
   return randomCards;
 }
-
-export const randomCardArray = () => randomArray(ICON_COUNT * 2);
-
-export const hasUnmatched = (cards) => cards.filter((c) => c.isFlipped && !c.isMatched).length === 2;
 
 export const clickCard = (cards, thisCard, clickCount) => {
 
@@ -50,9 +50,11 @@ export const clickCard = (cards, thisCard, clickCount) => {
       //exposedClicks : thisCard.exposedClicks.concat(clickCount),
       isFlipped: false, 
       isMatched: true }) ;
+
     const newCards = cards.map(c => c.name !== thisCard.name 
         ? c 
         : Object.assign({}, c.index === thisCard.index ? cc : c , {isFlipped: false, isMatched: true}));
+
     return {cards: newCards, clickedCard: cc };
   }
   
@@ -64,66 +66,4 @@ export const clickCard = (cards, thisCard, clickCount) => {
   });
 
   return {cards: cards.map(c => c.index === cc.index ? cc : c), clickedCard: cc };
-}
-
-
-//  this methosd is still a work in progress.  usage:  gameScore(state.memoryGame.history)
-export const gameScore = (clickHistory) => {
-  // TODO: compute how well this user performed based upon the click history
-
-  let matches = [];
-  for(let i = 0; i < clickHistory.length; i++) {
-    if(clickHistory[i].isMatched){
-      matches.push({
-          name: clickHistory[i].name,
-          first: clickHistory[i-1].exposedClicks.length, 
-          second: clickHistory[i].exposedClicks.length,
-          matchCount: matches.length
-        });
-    }
-  }
-
-  //  last match is easy
-  const lastMatch =  matches.filter(m => m.matchCount === 11);  
-  let remains = matches.filter(m => !lastMatch.map(p => p.name).includes(m.name));
-
-  // count the lucky matches, then remove them
-  const luck = remains.filter(m => m.second === 0); 
-  remains = remains.filter(m => !luck.map(p => p.name).includes(m.name));
-
-  // count the perfect matches, then remove them
-  const level1 = remains.filter(m => (m.second === 1 && m.first === 2) || (m.first === 1 && m.second === 2));
-  remains = remains.filter(m => !level1.map(p => p.name).includes(m.name));
-
-  //counts near perfect matches
-  const level2 = remains.filter(m => m.second < 2 || m.first < 2);
-  remains = remains.filter(m => !level2.map(p => p.name).includes(m.name));
-
-  // count matches where both cards seen a couple of times
-  const level3 = remains.filter(m => m.second < 3 && m.first < 3);
-
-  // count of all the matches where user turned both cards over serveral times
-  const level4 = remains = remains.filter(m => !level3.map(p => p.name).includes(m.name));
-
-  // what is the most times a card was clicked?
-  const maxSeen = Math.max(...matches.map(m => m.first), ...matches.map(m => m.second));
-
-  // which cards were clicked the most?
-  const flipCounts = matches.sort((a, b) => {
-    const ma = Math.max(a.first, a.second);
-    const mb = Math.max(b.first, b.second);
-
-    if (ma === mb ) {
-      const at = a.first + a.second;
-      const bt = b.first + b.second;
-      return  at === bt ? 0 : at < bt ? 1 : -1;
-    }
-
-    return ma < mb ? 1 : -1;
-  });
-
-  const result = {luck, level1, level2, level3, level4, lastMatch, maxSeen, flipCounts};
-
-  console.log('result:', maxSeen, flipCounts, result);
-  return result;
 }
